requirements.txt
aiohttp
aiosqlite
discord.py
pypresence